Make preview images/videos unselectable by keyboard
This was decided because the preview GIFs were made for people that could glance rapidly
through the projects. If somebody is using a screen reader, it would be very slow for the
software to describe meaninfully the GIFs. As all of them are already inside the modals of
the projects with proper description I thought the preview ones could make the experience worse
for people that use screens readears.